changing import locations for moved files
