+++ .gitignore
@@ -4,5 +4,4 @@ envVariables.js
 .idea
 *.log
 output.json
-dbStuff.js
 icon.png
